requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.3
requests==2.31.0
python-dotenv==1.0.0

sentence-transformers==5.2.0
faiss-cpu==1.8.0
numpy==1.26.4
networkx==3.3
torch==2.0.1
transformers==4.41.2
scikit-learn==1.4.2

pytest==7.4.4
